test: gardening
- Separate worker for some Android tests.
- Use png comprator for some screenshots tests instead of buffer equality.
- Various timeout fixes.
